Refactor project structure and cleanup
