+++ internal/common/apps.go
