+++ internal/common/basic.go
