+++ providers/capa/standard/capa_test.go
@@ -3,7 +3,7 @@ package standard
 import (
 	. "github.com/onsi/ginkgo/v2"
 
-	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/common"
 )
 
 var _ = Describe("Common tests", func() {
