+++ providers/capvcd/standard/capvcd_test.go
@@ -3,7 +3,7 @@ package standard
 import (
 	. "github.com/onsi/ginkgo/v2"
 
-	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/common"
 )
 
 var _ = PDescribe("Common tests", func() {
